libyang REFACTOR use size-specific int types instead of generic types
Align parameters' types with the values in (internal) structures.

+++ compat/compat.c
@@ -128,7 +128,7 @@ getline(char **lineptr, size_t *n, FILE *stream)
 {
     static char line[256];
     char *ptr;
-    unsigned int len;
+    ssize_t len;
 
     if (!lineptr || !n) {
         errno = EINVAL;
