+++ src/printer_schema.c
@@ -25,8 +25,8 @@
 #include "tree_schema.h"
 
 API LY_ERR
-lys_print_module(struct ly_out *out, const struct lys_module *module, LYS_OUTFORMAT format, int UNUSED(line_length),
-        int options)
+lys_print_module(struct ly_out *out, const struct lys_module *module, LYS_OUTFORMAT format, size_t UNUSED(line_length),
+        uint32_t options)
 {
     LY_ERR ret;
 
@@ -82,7 +82,7 @@ lys_print_module(struct ly_out *out, const struct lys_module *module, LYS_OUTFOR
 
 API LY_ERR
 lys_print_submodule(struct ly_out *out, const struct lys_module *module, const struct lysp_submodule *submodule,
-        LYS_OUTFORMAT format, int UNUSED(line_length), int options)
+        LYS_OUTFORMAT format, size_t UNUSED(line_length), uint32_t options)
 {
     LY_ERR ret;
 
@@ -116,7 +116,7 @@ lys_print_submodule(struct ly_out *out, const struct lys_module *module, const s
 }
 
 static LY_ERR
-lys_print_(struct ly_out *out, const struct lys_module *module, LYS_OUTFORMAT format, int options)
+lys_print_(struct ly_out *out, const struct lys_module *module, LYS_OUTFORMAT format, uint32_t options)
 {
     LY_ERR ret;
 
@@ -129,7 +129,7 @@ lys_print_(struct ly_out *out, const struct lys_module *module, LYS_OUTFORMAT fo
 }
 
 API LY_ERR
-lys_print_mem(char **strp, const struct lys_module *module, LYS_OUTFORMAT format, int options)
+lys_print_mem(char **strp, const struct lys_module *module, LYS_OUTFORMAT format, uint32_t options)
 {
     struct ly_out *out;
 
@@ -143,7 +143,7 @@ lys_print_mem(char **strp, const struct lys_module *module, LYS_OUTFORMAT format
 }
 
 API LY_ERR
-lys_print_fd(int fd, const struct lys_module *module, LYS_OUTFORMAT format, int options)
+lys_print_fd(int fd, const struct lys_module *module, LYS_OUTFORMAT format, uint32_t options)
 {
     struct ly_out *out;
 
@@ -154,7 +154,7 @@ lys_print_fd(int fd, const struct lys_module *module, LYS_OUTFORMAT format, int 
 }
 
 API LY_ERR
-lys_print_file(FILE *f, const struct lys_module *module, LYS_OUTFORMAT format, int options)
+lys_print_file(FILE *f, const struct lys_module *module, LYS_OUTFORMAT format, uint32_t options)
 {
     struct ly_out *out;
 
@@ -165,7 +165,7 @@ lys_print_file(FILE *f, const struct lys_module *module, LYS_OUTFORMAT format, i
 }
 
 API LY_ERR
-lys_print_path(const char *path, const struct lys_module *module, LYS_OUTFORMAT format, int options)
+lys_print_path(const char *path, const struct lys_module *module, LYS_OUTFORMAT format, uint32_t options)
 {
     struct ly_out *out;
 
@@ -176,7 +176,7 @@ lys_print_path(const char *path, const struct lys_module *module, LYS_OUTFORMAT 
 }
 
 API LY_ERR
-lys_print_clb(ly_write_clb writeclb, void *user_data, const struct lys_module *module, LYS_OUTFORMAT format, int options)
+lys_print_clb(ly_write_clb writeclb, void *user_data, const struct lys_module *module, LYS_OUTFORMAT format, uint32_t options)
 {
     struct ly_out *out;
 
@@ -187,7 +187,7 @@ lys_print_clb(ly_write_clb writeclb, void *user_data, const struct lys_module *m
 }
 
 API LY_ERR
-lys_print_node(struct ly_out *out, const struct lysc_node *node, LYS_OUTFORMAT format, int UNUSED(line_length), int options)
+lys_print_node(struct ly_out *out, const struct lysc_node *node, LYS_OUTFORMAT format, size_t UNUSED(line_length), uint32_t options)
 {
     LY_ERR ret;
